YT-18571: Preliminary cosmetics and refactoring of log compression commit_hash:57f52a44eb68d2c600e48de1c1234d6b708e4472
