+++ yt/yt/core/logging/file_log_writer.cpp
@@ -1,5 +1,6 @@
 #include "file_log_writer.h"
 
+#include "appendable_compressed_file.h"
 #include "log_writer_detail.h"
 #include "config.h"
 #include "private.h"
@@ -7,9 +8,8 @@
 #include "random_access_gzip.h"
 #include "stream_output.h"
 #include "system_log_event_provider.h"
-#include "compression.h"
 #include "log_writer_factory.h"
-#include "zstd_compression.h"
+#include "zstd_log_codec.h"
 #include "formatter.h"
 
 #include <yt/yt/core/misc/fs.h>
@@ -186,11 +186,13 @@ private:
             if (Config_->EnableCompression) {
                 switch (Config_->CompressionMethod) {
                     case ECompressionMethod::Zstd:
-                        OutputStream_ = New<TAppendableCompressedFile>(
+                        OutputStream_ = CreateAppendableCompressedFile(
                             *File_,
-                            CreateZstdCompressionCodec(Config_->CompressionLevel),
+                            CreateZstdLogCodec(Config_->CompressionLevel),
                             Host_->GetCompressionInvoker(),
-                            /*writeTruncateMessage*/ true);
+                            {
+                                .WriteTruncateMessage = true,
+                            });
                         break;
 
                     case ECompressionMethod::Gzip:
