+++ yt/yt/core/logging/zstd_log_codec.h
@@ -11,7 +11,7 @@ namespace NYT::NLogging {
 constexpr i64 MaxZstdFrameUncompressedLength = 5_MBs;
 constexpr const int DefaultZstdCompressionLevel = 3;
 
-ILogCompressionCodecPtr CreateZstdCompressionCodec(int compressionLevel = DefaultZstdCompressionLevel);
+ILogCodecPtr CreateZstdLogCodec(int compressionLevel = DefaultZstdCompressionLevel);
 
 ////////////////////////////////////////////////////////////////////////////////
